build: enable strict warnings and fix all warnings in project sources
- Add -Wall -Wextra -Wpedantic and conformance-focused flags to CMakeLists.txt,
  applied after submodules/ so third-party targets are unaffected
- Suppress warnings from DeckLink SDK (SYSTEM includes + -w on target)
  and stb (SYSTEM includes + -w on target)
- Fix -Wreorder: reorder constructor initializer lists in app_state.cpp
  and decklink/output.cpp to match member declaration order
- Fix -Wshadow: rename 'id' structured bindings to 'iface_id' in
  node_manager.cpp, rename inner 'con_it' to 'existing_it', remove
  unused shadowing 'state' variable in interface.cpp
- Fix -Wold-style-cast: replace C casts with static_cast/reinterpret_cast
  in color_transfer.hpp (x2) and decklink_ptr.hpp
- Fix -Wdouble-promotion: use 1.0f instead of 1.0 in color_transfer.hpp
- Fix -Wunused-variable: [[maybe_unused]] on has_dvp/has_amd_pinned in
  transfer.cpp; remove dead 'dim' variable in output.cpp constructor
- Fix -Wunused-parameter: comment out parameter names in ffmpeg/player.cpp
- Fix -Wsign-compare: add static_cast<size_t/int> in interface.cpp,
  teleprompter.cpp, font_instance.cpp; use size_t loop var in bundler
- Fix -Wunused-function: change static to inline in flicks.hpp and
  thread_priority.hpp
- Fix -Wpessimizing-move: remove redundant std::move on future::get()
  return value in teleprompter.cpp

+++ src/render/font/font_instance.cpp
@@ -41,7 +41,7 @@ font_instance_s::flow_info_s font_instance_s::flow_line(std::u32string_view str,
         const char32_t c = str[i];
 
         if (std::iswspace(c) != 0 || i == str.size() - 1) {
-            if (info.pixels_advanced + word_len < width) {
+            if (info.pixels_advanced + word_len < static_cast<size_t>(width)) {
                 info.pixels_advanced += word_len;
                 word_len            = 0;
                 info.consumed_chars = i + 1;
